requirements.txt
click==8.1.3
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
jdatetime==4.1.1
numpy==1.25.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
Pillow==9.5.0
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
